// types.ts
export interface Theme {
  name: string;
  background: string;
  clockBackground: string;
  borderColor: string;
  textColor: string;
  colonColor: string;
  dateColor: string;
  buttonBg: string;
  buttonColor: string;
  buttonBorder: string;
}

export interface Font {
  name: string;
  styleName: string;
}

export interface MusicTrack {
  title: string;
  url: string;
}